USB KEYBOARD/MOUSE: added on/off trigger to driver. READ MORE [dorimanx]
- 03 September 2014 ported (fixing compile error) for Samsung Galaxy Tab S SM-705 LTE by UpInTheAir@XDA

When was ON by default I have seen that USB is always HID device.

And i have always received 2 connection sounds and vibrate...

It's annoying! so now all good, one alert for connection, and USB is Android Device, and when I have connect USB Keyboard or mouse or both via wireless connector (Microsoft keyboard + mouse Infra-Red) it works! Phone detect mouse + keyboard and I can use them on screen!

The on/off trigger seems not needed to use the HID and can be OFF. I have left it for future use, for compatibility if needed.